fix: consolidate MCP runtime tests into execution_test.go for pkg-file-count
Fold runtime-backed MCP client tests into execution_test.go so
pkg/mcp/factorysession stays within the 15-file package limit.
